reorganize in preparation for more adapters
